requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.8
django-filter==2.3.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
drf-yasg==1.17.1
idna==2.10
inflection==0.5.0
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.4
Pillow==7.2.0
psycopg2==2.8.5
PyJWT==1.7.1
pylint==2.5.3
pyparsing==2.4.7
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
wrapt==1.12.1
